Adding TLS verify support & key.pem support.
With this change, the cmdlets will be able to use the ca.pem/key.pem combination in the same way as the docker CLI, removing the need for generating a separate, Windows-specific PFX.  Unforunately, the mechanisms required to make this work are not yet available in .NET Core, so certificate support is Windows only at the moment.  Once .NET Core with the latest networking fixes is released, we can unblock that scenario for cross-platform.